+++ io/silvicky/novel/compiler/code/AssignNumberCode.java
@@ -2,7 +2,7 @@ package io.silvicky.novel.compiler.code;
 
 import static io.silvicky.novel.compiler.Compiler.lookupVariableName;
 
-public record AssignNumberCode(long target, long left) implements Code
+public record AssignNumberCode(int target, long left) implements Code
 {
     public String toString()
     {
