+++ io/silvicky/novel/compiler/code/LabelCode.java
@@ -3,10 +3,10 @@ package io.silvicky.novel.compiler.code;
 import static io.silvicky.novel.compiler.Compiler.lookupLabelName;
 import static io.silvicky.novel.compiler.Compiler.requestInternalLabel;
 
-public record LabelCode(long id) implements Code
+public record LabelCode(int id) implements Code
 {
     public LabelCode(){this(requestInternalLabel());}
-    public LabelCode(long id){this.id=id;}
+    public LabelCode(int id){this.id=id;}
 
     @Override
     public String toString()
