+++ io/silvicky/novel/compiler/code/UnconditionalGotoCode.java
@@ -4,7 +4,7 @@ import io.silvicky.novel.compiler.tokens.OperatorType;
 
 import static io.silvicky.novel.compiler.Compiler.lookupLabelName;
 
-public record UnconditionalGotoCode(long id) implements Code
+public record UnconditionalGotoCode(int id) implements Code
 {
     @Override
     public String toString()
